requirements.txt
feedparser>=6.0.0
pymongo>=4.0.0
certifi>=2024.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
trafilatura>=2.0.0
ollama>=0.6.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
slowapi>=0.1.9
